% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtop4.R
\docType{data}
\name{brazil}
\alias{brazil}
\title{COVID19 dataset for Brazil}
\format{
A dataframe of 5944 rows and 8 variables
\itemize{
\item \strong{Date}: The date at which the count of confirmed  cases is recorded.
\item \strong{state}: Abbreviation of the states within the country for which the covid pandemic was affected.
\item \strong{deaths}: The number of people who have been recorded for deaths by the virus.
\item \strong{cases}: Number of cases recorded within the state.
\item \strong{state_name}: Names of the states in USA .
\item \strong{lat}: The latitude of the state.
\item \strong{long}: The longitude of the state.
}
}
\usage{
brazil
}
\description{
A dataset for the number of coronavirus cases within the country Brazil, cases, deaths and recovered.Data available till October 2020.
}
\references{
\href{https://www.kaggle.com/unanimad/corona-virus-brazil}{Brazil}
}
\keyword{datasets}
