% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtop4.R
\docType{data}
\name{corona}
\alias{corona}
\title{COVID19 dataset real-time}
\format{
A dataframe with 271 rows and 35 variables:
\itemize{
\item \strong{id}: The 3 character country ID for every country in the world.
\item \strong{date}: The date for each day the cases were recorded
\item \strong{test}: The number of tests conducted in that day in that country.
\item \strong{confirmed}:The number of cases that were tested to be positive.
\item \strong{recovered}: The number of cases that were recovered that day in the country.
\item \strong{deaths}: The number of deaths that day in the country.
\item \strong{population}: The number of people within that country(census).
\item \strong{school closing}: The number of schools closed during the pandemic each day.
}
}
\usage{
corona
}
\description{
A real time dataset which contains the coronavirus cases, deaths and recovered within each country throughout the world each day.
}
\references{
\href{https://cran.r-project.org/package=COVID19}{COVID19 package}
}
\keyword{datasets}
