% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtop4.R
\docType{data}
\name{india}
\alias{india}
\title{COVID19 dataset for India}
\format{
A dataframe of 6806 rows and 11 variables
\itemize{
\item \strong{Date}: The date at which the count of confirmed  cases is recorded.
\item \strong{Time}: The time at which the cases was recorded.
\item \strong{states}: Name of the states within the country for which the covid pandemic was affected.
\item \strong{ConfrmedIndiaNational}: The number of Indian nationals who are recorded to have been affected.
\item \strong{ConfirmedForeignNational}: The number of foreign nationals who are recorded to have been affected with the pandemic.
\item \strong{Cured}: The number of people who have been recovered on that day in that state.
\item \strong{Deaths}: The number of people who have been recorded for deaths by the virus.
\item \strong{Confirmed}: Number of cases recorded within the state.
\item \strong{lat}: The latitude of the state.
\item \strong{long}: The longitude of the state.
}
}
\usage{
india
}
\description{
A dataset for the number of coronavirus cases within the country India, cases, deaths and recovered.Data available till October 2020.
}
\references{
\href{https://www.kaggle.com/sudalairajkumar/covid19-in-india}{India}
}
\keyword{datasets}
