% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_maps.R
\name{leaflet_maps}
\alias{leaflet_maps}
\title{Leaflet maps for each country}
\usage{
leaflet_maps(longitude, latitude, state, total_cases)
}
\arguments{
\item{longitude}{A column passed from the dataframe which contains the longitude for the maps.}

\item{latitude}{A column passed from the dataframe which contains the latitude for the maps.}

\item{state}{A column passed from the dataframe which contains the state names for the popup on the leaflet maps.}

\item{total_cases}{A column passed from the dataframe which cintains the total number of cases for each state.}
}
\value{
returns a leaflet map with longitudes and latitudes for the map with state names.
}
\description{
This function helps the user make a leaflet map using the latitude and longitudes for the states/region/country you are visualizing. There are many datasets within this package which can used as examples for checking the function like the BRA_distinct for Brazilian states, IND_distinct for Indian states, USA_distinct  for USA states and RUS_distinct for Russian states.
}
\examples{
\dontrun{
leaflet(longitude = df$lng,latitude = df$lat, state = df$state,total_cases = df$total_cases)
}

}
\author{
Varsha Ujjinni Vijay Kumar
}
