% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtop4.R
\docType{data}
\name{RUS_distinct}
\alias{RUS_distinct}
\title{COVID19 dataset for Russian individual states}
\format{
A dataframe of 84 rows and  4 columns
\itemize{
\item \strong{state}: The name of the states and territories
\item \strong{lat}: The latitude of the state
\item \strong{long}: The longitude of the state
\item \strong{total}: The total number of cases within each state
}
}
\usage{
RUS_distinct
}
\description{
A dataset for the number of coronavirus cases within the Russian states and territories.
}
\references{
taken from the above websites and some data acquiring from wikipedia from my end
}
\keyword{datasets}
