% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtop4.R
\docType{data}
\name{russia}
\alias{russia}
\title{COVID19 dataset for Russia}
\format{
A dataframe of 61209 rows and 8 variables
\itemize{
\item \strong{Date.x}: The date at which the count of confirmed  cases is recorded.
\item \strong{Deaths}: The number of people who have been recorded for deaths by the virus.
\item \strong{Confirmed}: Number of cases recorded within the state.
\item \strong{Recovered}: Number of people who have are recovered that day.
\item \strong{state}: Names of the states in USA .
\item \strong{lat}: The latitude of the state.
\item \strong{long}: The longitude of the state.
}
}
\usage{
russia
}
\description{
A dataset for the number of coronavirus cases within the country Russia, cases, deaths and recovered.Data available till October 2020.
}
\references{
\href{https://www.kaggle.com/kapral42/covid19-russia-regions-cases}{Russia}
}
\keyword{datasets}
