% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateInput.R
\name{stateInput}
\alias{stateInput}
\title{State Select Input}
\usage{
stateInput(id, x)
}
\arguments{
\item{id:}{The input ID of the select Input that is passed to the app by the user.}

\item{x:}{The column in the dataframe that contains the country names.}
}
\description{
Creates a shiny select input to select a state from a column of the dataset
}
\details{
The purpose of this function is to provide a quick select input for different values in the column.
}
\examples{
\dontrun{
stateInput(id = "state", x = usa$state)
}

}
\author{
Varsha Ujjinni Vijay Kumar
}
