% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtop4.R
\docType{data}
\name{usa}
\alias{usa}
\title{COVID19 dataset for United States of America}
\format{
A dataframe of 11634 rows and 11 variables
\itemize{
\item \strong{Date}: The date at which the count of confirmed  cases is recorded.
\item \strong{state}: Abbreviation of the states within the country for which the covid pandemic was affected.
\item \strong{recovered}: The number of people who have been recovered on that day in that state.
\item \strong{death}: The number of people who have been recorded for deaths by the virus.
\item \strong{positive}: Number of cases recorded within the state.
\item \strong{total}: Total number of cases within each state.
\item \strong{state_name}: Names of the states in USA .
\item \strong{lat}: The latitude of the state.
\item \strong{long}: The longitude of the state.
}
}
\usage{
usa
}
\description{
A dataset for the number of coronavirus cases within the country USA, cases, deaths and recovered.Data available till October 2020.
}
\references{
\href{https://www.kaggle.com/sudalairajkumar/covid19-in-usa}{USA}
}
\keyword{datasets}
