% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtop4.R
\docType{data}
\name{USA_distinct}
\alias{USA_distinct}
\title{COVID19 dataset for USA individual states}
\format{
A dataframe of 54 rows and  4 columns
\itemize{
\item \strong{state_name}: The name of the states and territories
\item \strong{lat}: The latitude of the state
\item \strong{lng}: The longitude of the state
\item \strong{total}: The total number of cases within each state
}
}
\usage{
USA_distinct
}
\description{
A dataset for the number of coronavirus cases within the USA states and territories.
}
\references{
taken from the above kaggle websites and some data acquiring from wikipedia from my end
}
\keyword{datasets}
