% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19top4-package.R
\docType{package}
\name{COVID19top4-package}
\alias{COVID19top4}
\alias{COVID19top4-package}
\title{COVID19top4: A Shiny App for Top 4 COVID19 Affected Countries}
\description{
It helps in containing a shiny application which launches COVID19 pandemic situation in the top four most affected countries as of October, 2020. This contains functions which help launch and use the app for information and knowledge on the pandemic during the time. A time graph to the world map for visualization is embeded into the application.
}
\author{
\strong{Maintainer}: Varsha Ujjinni Vijay Kumar \email{varshaujjinnivijaykumar@gmail.com}

}
\keyword{internal}
