% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_maps.R
\name{leaflet_maps}
\alias{leaflet_maps}
\title{Leaflet maps for each country}
\usage{
BRA_distinct

IND_distinct

USA_distinct

RUS_distinct
}
\arguments{
\item{data_frame:}{A dataframe which is passed for the leaflet to be produced from}

\item{longitude:}{A column passed from the dataframe which contains the longitude for the maps.}

\item{latitude:}{A column passed from the dataframe which contains the latitude for the maps.}

\item{state:}{A column passed from the dataframe which contains the state names for the popup on the leaflet maps.}

\item{total_cases:}{A column passed from the dataframe which cintains the total number of cases for each state.}
}
\value{
returns a leaflet map with longitudes and latitudes for the map with state names.
}
\description{
Leaflet maps for each country
}
\examples{
\dontrun{
leaflet(data_frame = df, longitude = df$lng,latitude = df$lat, state = df$state,total = df$total_cases)
}


}
\author{
Varsha Ujjinni Vijay Kumar
}
